Move MCP reachability checks into unit coverage
