+++ tests/e2e/test_mcp_e2e.py
@@ -21,6 +21,8 @@ from mcp import ClientSession
 from mcp.client.streamable_http import streamable_http_client
 from mcp.types import CallToolResult, ListToolsResult, TextContent
 
+from slcli.mcp_reachability import is_reachability_failure
+
 DEFAULT_MCP_URL = "http://127.0.0.1:8000/mcp"
 DEFAULT_TIMEOUT_SECONDS = 5
 MISSING_RESOURCE_SENTINEL = "mcp-e2e-missing-resource"
@@ -157,70 +159,6 @@ def _comment_target(state: Dict[str, Any]) -> Tuple[str, str]:
             return resource_type, resource_id
 
     return "workitem:workitem", MISSING_RESOURCE_SENTINEL
-
-
-def _iter_exceptions(exc: BaseException) -> Iterator[BaseException]:
-    """Yield an exception plus any nested causes, contexts, or exception-group members."""
-    stack: List[BaseException] = [exc]
-    seen: set[int] = set()
-
-    while stack:
-        current = stack.pop()
-        marker = id(current)
-        if marker in seen:
-            continue
-        seen.add(marker)
-        yield current
-
-        cause = getattr(current, "__cause__", None)
-        if isinstance(cause, BaseException):
-            stack.append(cause)
-
-        context = getattr(current, "__context__", None)
-        if isinstance(context, BaseException):
-            stack.append(context)
-
-        if isinstance(current, BaseExceptionGroup):
-            stack.extend(current.exceptions)
-
-
-def _is_reachability_failure(exc: Exception) -> bool:
-    """Return True when an exception indicates the local MCP server is unreachable."""
-    for candidate in _iter_exceptions(exc):
-        if isinstance(candidate, (OSError, TimeoutError)):
-            return True
-
-        message = str(candidate).lower()
-        if any(
-            token in message
-            for token in (
-                "connection refused",
-                "connect error",
-                "all connection attempts failed",
-                "timed out",
-                "timeout",
-                "name or service not known",
-                "nodename nor servname provided",
-                "server disconnected",
-            )
-        ):
-            return True
-
-    return False
-
-
-def test_is_reachability_failure_detects_nested_exception_group() -> None:
-    """Nested connection failures should still be treated as an unreachable local server."""
-    exc = ExceptionGroup("task group failure", [OSError("Connection refused")])
-
-    assert _is_reachability_failure(exc) is True
-
-
-def test_is_reachability_failure_rejects_unrelated_exception_group() -> None:
-    """Unrelated exception groups should not be mistaken for local reachability failures."""
-    exc = ExceptionGroup("task group failure", [RuntimeError("boom")])
-
-    assert _is_reachability_failure(exc) is False
 
 
 async def _call_tool(
@@ -516,7 +454,7 @@ async def _exercise_mcp_tools(mcp_url: str, timeout_seconds: int) -> None:
                         allow_error=True,
                     )
     except Exception as exc:
-        if _is_reachability_failure(exc):
+        if is_reachability_failure(exc):
             pytest.skip(f"Local MCP server is not reachable at {mcp_url}: {exc}")
         raise
